fix: resolve CI web startup graph budget regression
